requirements.txt
requests
Pillow
pandas
google-generativeai
python-dotenv
